Adding clean up of records based on the model schema to save and update records.
